Fixes #172
I'm not sure if this is an actual fix, but it's difficult to recreate.
This allows for multiple writers and readers, so there may be some
instances where the highlighted cells are incorrect, but I'm not sure
how that would happen. This also should only happen when rapidly
visualizing bots, so shouldn't ever be noticeable anyway.