Fix resizing the window breaking things: see #131
This PR fixes #130. It has exactly the same fix as PR #131, but doesn't
include other changes to the json files and such.

@dudecon: pro tip! Learn how to `git rebase`. It lets you re-write
history, only allow certain parts of commits, etc... Also thanks for the
code for the fix -- I would have no idea how to do that. You're the
MVP here.