Add member IsWhiteToMoveInt
Whilst the `IsWhiteToMove` can be used in boolean logic to determin the colour that is moving, math cannot. `IsWhiteToMoveInt` will fix this by enabling bots to get a number they can multiply against.

For example, when using a numerical evaluation that has white is positive, the bot can multiply against `IsWhiteToMoveInt` at each step in the search tree to find if the evaluation is better for the current player or opponent.